Rename "default_cpu" to "default"
The default target is not compatible only with CPU, so renaming files
and references from "default_cpu" to "default"
